refactor: move export logic to separate module
